+++ src/ts/MRAID/AdUnits/MRAIDAdUnitFactory.ts
@@ -3,33 +3,38 @@ import { CustomFeatures } from 'Ads/Utilities/CustomFeatures';
 import { AbstractPrivacy } from 'Ads/Views/AbstractPrivacy';
 import { ARMRAID } from 'AR/Views/ARMRAID';
 import { IMRAIDAdUnitParameters, MRAIDAdUnit } from 'MRAID/AdUnits/MRAIDAdUnit';
-import { MRAIDEventHandler } from 'MRAID/EventHandlers/MRAIDEventHandler';
-import { PlayableEventHandler } from 'MRAID/EventHandlers/PlayableEventHandler';
 import { MRAIDCampaign } from 'MRAID/Models/MRAIDCampaign';
 import { IMRAIDViewHandler } from 'MRAID/Views/MRAIDView';
 import { PerformanceMRAIDCampaign } from 'Performance/Models/PerformanceMRAIDCampaign';
+import { PerformanceMRAIDEventHandler } from 'MRAID/EventHandlers/PerformanceMRAIDEventHandler';
 import { ARMRAIDEventHandler } from 'AR/EventHandlers/ARMRAIDEventHandler';
-import { WebPlayerMRAIDAdUnit } from 'MRAID/AdUnits/WebPlayerMRAIDAdUnit';
+import { ProgrammaticMRAIDEventHandler } from 'MRAID/EventHandlers/ProgrammaticMRAIDEventHandler';
 import { WebPlayerMRAIDTest } from 'Core/Models/ABGroup';
+import { WebPlayerMRAIDAdUnit } from 'MRAID/AdUnits/WebPlayerMRAIDAdUnit';
 
 export class MRAIDAdUnitFactory extends AbstractAdUnitFactory<MRAIDCampaign, IMRAIDAdUnitParameters> {
     public createAdUnit(parameters: IMRAIDAdUnitParameters): MRAIDAdUnit {
-        let mraidAdUnit;
+        let mraidAdUnit: MRAIDAdUnit = new MRAIDAdUnit(parameters);
+
         if (WebPlayerMRAIDTest.isValid(parameters.coreConfig.getAbGroup())) {
             mraidAdUnit = new WebPlayerMRAIDAdUnit(parameters);
         } else {
             mraidAdUnit = new MRAIDAdUnit(parameters);
         }
 
-        // NOTE: When content type is correct for playables we want to change this to content type check.
-        const isPlayable: boolean = parameters.campaign instanceof PerformanceMRAIDCampaign;
-        const isAR: boolean = parameters.mraid instanceof ARMRAID;
-        const isSonicPlayable: boolean = CustomFeatures.isSonicPlayable(parameters.campaign.getCreativeId());
-        const EventHandler = (isSonicPlayable || isPlayable) ? PlayableEventHandler :
-            isAR ? ARMRAIDEventHandler : MRAIDEventHandler;
-        const mraidEventHandler: IMRAIDViewHandler = new EventHandler(mraidAdUnit, parameters);
+        const mraidEventHandler: IMRAIDViewHandler = this.getMRAIDEventHandler(mraidAdUnit, parameters);
         parameters.mraid.addEventHandler(mraidEventHandler);
         AbstractPrivacy.setupReportListener(parameters.privacy, mraidAdUnit);
         return mraidAdUnit;
+    }
+
+    private getMRAIDEventHandler(mraidAdUnit: MRAIDAdUnit, parameters: IMRAIDAdUnitParameters): IMRAIDViewHandler {
+        if (CustomFeatures.isSonicPlayable(parameters.campaign.getCreativeId()) || parameters.campaign instanceof PerformanceMRAIDCampaign) {
+            return new PerformanceMRAIDEventHandler(mraidAdUnit, parameters);
+        } else if (parameters.mraid instanceof ARMRAID) {
+            return new ARMRAIDEventHandler(mraidAdUnit, parameters);
+        } else {
+            return new ProgrammaticMRAIDEventHandler(mraidAdUnit, parameters);
+        }
     }
 }